requirements.txt
# installs dependencies from ./data-tools/setup.py, and the package itself,
-e ./data/data-tools/
-e ./data/fhg-utils/

jupyter
ipython
matplotlib
seaborn
tabulate
numpy
pytorch-tcn
librosa
imbalanced-learn
